more

+++ src/Element.java
@@ -31,12 +31,12 @@ public abstract class Element{
 	    return false;
 	}
 
-	public void setX(int x) {
-		this.x = x;
+	public void setX(double x) {
+		this.x = (int)x;
 	}
 	
-	public void setY(int y) {
-		this.y = y;
+	public void setY(double y) {
+		this.y = (int)y;
 	}
 	
 	public int getX() {
